# openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Agent Orcha API",
        "version": "0.0.3",
        "description": "API for orchestrating agents, workflows, knowledge stores, functions, MCP servers, and LLMs."
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/workflows/{name}": {
            "get": {
                "summary": "Get Workflow Details",
                "tags": ["Workflows"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Workflow details",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    },
                    "404": { "description": "Workflow not found" }
                }
            }
        },
        "/api/workflows/{name}/run": {
            "post": {
                "summary": "Run Workflow",
                "tags": ["Workflows"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "input": { "type": "object" }
                                },
                                "required": ["input"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Workflow execution result",
                        "content": {
                            "application/json": {
                                "schema": { "type": "object" }
                            }
                        }
                    },
                    "404": { "description": "Workflow not found" },
                    "500": { "description": "Internal server error" }
                }
            }
        },
        "/api/workflows/{name}/stream": {
            "post": {
                "summary": "Stream Workflow Execution",
                "tags": ["Workflows"],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "string" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "input": { "type": "object" }
                                },
                                "required": ["input"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Server-Sent Events stream",
                        "content": {
                            "text/event-stream": {
                                "schema": { "type": "string" }
                            }
                        }
                    }
                }
            }
        },
...
    }
}